+++ lib/src/webos_m_tokens/effect_semantic_dark_res_low.dart
@@ -24,6 +24,4 @@ class LowEffectSemantic {
   int get pageIndicatorDropShadowBlur => EffectPrimitive.instance.effect8;
   int get pageIndicatorDropShadowY => EffectPrimitive.instance.effect1;
   int get statusBarBackgroundBlur => EffectPrimitive.instance.effect0;
-  int get widgetDropShadowBlur => EffectPrimitive.instance.effect40;
-  int get widgetDropShadowY => EffectPrimitive.instance.effect40;
 }